+++ integration/query_frontend_cache_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/grafana/dskit/cache"
 	"github.com/grafana/e2e"
 	e2ecache "github.com/grafana/e2e/cache"
 	e2edb "github.com/grafana/e2e/db"
@@ -17,70 +18,73 @@ import (
 )
 
 func TestQueryFrontendUnalignedQuery(t *testing.T) {
-	s, err := e2e.NewScenario(networkName)
-	require.NoError(t, err)
-	defer s.Close()
+	for _, backend := range []string{cache.BackendMemcached, cache.BackendRedis} {
+		t.Run(backend, func(t *testing.T) {
+			s, err := e2e.NewScenario(networkName)
+			require.NoError(t, err)
+			defer s.Close()
 
-	memcached := e2ecache.NewMemcached()
-	consul := e2edb.NewConsul()
-	require.NoError(t, s.StartAndWaitReady(consul, memcached))
+			cacheService := cacheBackend(backend)
+			consul := e2edb.NewConsul()
+			require.NoError(t, s.StartAndWaitReady(consul, cacheService))
 
-	const configFile = ""
-	flags := mergeFlags(
-		BlocksStorageFlags(),
-		BlocksStorageS3Flags(),
-	)
+			const configFile = ""
+			flags := mergeFlags(
+				BlocksStorageFlags(),
+				BlocksStorageS3Flags(),
+			)
 
-	minio := e2edb.NewMinio(9000, flags["-blocks-storage.s3.bucket-name"])
-	require.NoError(t, s.StartAndWaitReady(minio))
+			minio := e2edb.NewMinio(9000, flags["-blocks-storage.s3.bucket-name"])
+			require.NoError(t, s.StartAndWaitReady(minio))
 
-	flags = mergeFlags(flags, map[string]string{
-		"-query-frontend.cache-results":                     "true",
-		"-query-frontend.split-queries-by-interval":         "2m",
-		"-query-frontend.align-queries-with-step":           "true",
-		"-query-frontend.max-cache-freshness":               "0", // Cache everything.
-		"-query-frontend.results-cache.backend":             "memcached",
-		"-query-frontend.results-cache.memcached.addresses": "dns+" + memcached.NetworkEndpoint(e2ecache.MemcachedPort),
+			flags = mergeFlags(flags, map[string]string{
+				"-query-frontend.cache-results":             "true",
+				"-query-frontend.split-queries-by-interval": "2m",
+				"-query-frontend.max-cache-freshness":       "0", // Cache everything.
+				// Enable protobuf format so that we can use native histograms.
+				"-query-frontend.query-result-response-format": "protobuf",
+			}, cacheConfig(backend, cacheService))
 
-		// Enable protobuf format so that we can use native histograms.
-		"-query-frontend.query-result-response-format": "protobuf",
-	})
+			// Start the query-frontend.
+			queryFrontendAligned := e2emimir.NewQueryFrontend("query-frontend-aligned", mergeFlags(flags, map[string]string{"-query-frontend.align-queries-with-step": "true"}), e2emimir.WithConfigFile(configFile))
+			require.NoError(t, s.Start(queryFrontendAligned))
 
-	// Start the query-frontend.
-	queryFrontendAligned := e2emimir.NewQueryFrontend("query-frontend-aligned", mergeFlags(flags, map[string]string{"-query-frontend.align-queries-with-step": "true"}), e2emimir.WithConfigFile(configFile))
-	require.NoError(t, s.Start(queryFrontendAligned))
+			queryFrontendUnaligned := e2emimir.NewQueryFrontend("query-frontend-unaligned", mergeFlags(flags, map[string]string{"-query-frontend.align-queries-with-step": "false"}), e2emimir.WithConfigFile(configFile))
+			require.NoError(t, s.Start(queryFrontendUnaligned))
 
-	queryFrontendUnaligned := e2emimir.NewQueryFrontend("query-frontend-unaligned", mergeFlags(flags, map[string]string{"-query-frontend.align-queries-with-step": "false"}), e2emimir.WithConfigFile(configFile))
-	require.NoError(t, s.Start(queryFrontendUnaligned))
+			querierAligned := e2emimir.NewQuerier("querier-aligned", consul.NetworkHTTPEndpoint(), mergeFlags(flags, map[string]string{"-querier.frontend-address": queryFrontendAligned.NetworkGRPCEndpoint()}), e2emimir.WithConfigFile(configFile))
+			querierUnaligned := e2emimir.NewQuerier("querier-unaligned", consul.NetworkHTTPEndpoint(), mergeFlags(flags, map[string]string{"-querier.frontend-address": queryFrontendUnaligned.NetworkGRPCEndpoint()}), e2emimir.WithConfigFile(configFile))
 
-	querierAligned := e2emimir.NewQuerier("querier-aligned", consul.NetworkHTTPEndpoint(), mergeFlags(flags, map[string]string{"-querier.frontend-address": queryFrontendAligned.NetworkGRPCEndpoint()}), e2emimir.WithConfigFile(configFile))
-	querierUnaligned := e2emimir.NewQuerier("querier-unaligned", consul.NetworkHTTPEndpoint(), mergeFlags(flags, map[string]string{"-querier.frontend-address": queryFrontendUnaligned.NetworkGRPCEndpoint()}), e2emimir.WithConfigFile(configFile))
+			// Start all other services.
+			ingester := e2emimir.NewIngester("ingester", consul.NetworkHTTPEndpoint(), flags, e2emimir.WithConfigFile(configFile))
+			distributor := e2emimir.NewDistributor("distributor", consul.NetworkHTTPEndpoint(), flags, e2emimir.WithConfigFile(configFile))
 
-	// Start all other services.
-	ingester := e2emimir.NewIngester("ingester", consul.NetworkHTTPEndpoint(), flags, e2emimir.WithConfigFile(configFile))
-	distributor := e2emimir.NewDistributor("distributor", consul.NetworkHTTPEndpoint(), flags, e2emimir.WithConfigFile(configFile))
+			require.NoError(t, s.StartAndWaitReady(querierAligned, querierUnaligned, ingester, distributor))
+			require.NoError(t, s.WaitReady(queryFrontendAligned, queryFrontendUnaligned))
 
-	require.NoError(t, s.StartAndWaitReady(querierAligned, querierUnaligned, ingester, distributor))
-	require.NoError(t, s.WaitReady(queryFrontendAligned, queryFrontendUnaligned))
+			// Check if we're discovering the cache only when running Memcached. The Redis client handles
+			// discovery itself when running a cluster or connects to a single machine.
+			if backend == cache.BackendMemcached {
+				require.NoError(t, queryFrontendAligned.WaitSumMetrics(e2e.Equals(1), "thanos_cache_dns_provider_results"))
+				require.NoError(t, queryFrontendUnaligned.WaitSumMetrics(e2e.Equals(1), "thanos_cache_dns_provider_results"))
+			}
 
-	// Check if we're discovering memcache or not.
-	require.NoError(t, queryFrontendAligned.WaitSumMetrics(e2e.Equals(1), "thanos_memcached_dns_provider_results"))
-	require.NoError(t, queryFrontendUnaligned.WaitSumMetrics(e2e.Equals(1), "thanos_memcached_dns_provider_results"))
+			// Wait until the distributor and queriers have updated the ring.
+			// The distributor should have 512 tokens for the ingester ring and 1 for the distributor ring
+			require.NoError(t, distributor.WaitSumMetrics(e2e.Equals(512+1), "cortex_ring_tokens_total"))
+			require.NoError(t, querierAligned.WaitSumMetrics(e2e.Equals(512), "cortex_ring_tokens_total"))
+			require.NoError(t, querierUnaligned.WaitSumMetrics(e2e.Equals(512), "cortex_ring_tokens_total"))
 
-	// Wait until the distributor and queriers have updated the ring.
-	// The distributor should have 512 tokens for the ingester ring and 1 for the distributor ring
-	require.NoError(t, distributor.WaitSumMetrics(e2e.Equals(512+1), "cortex_ring_tokens_total"))
-	require.NoError(t, querierAligned.WaitSumMetrics(e2e.Equals(512), "cortex_ring_tokens_total"))
-	require.NoError(t, querierUnaligned.WaitSumMetrics(e2e.Equals(512), "cortex_ring_tokens_total"))
+			// Push a series for each user to Mimir.
+			const user = "user"
 
-	// Push a series for each user to Mimir.
-	const user = "user"
+			c, err := e2emimir.NewClient(distributor.HTTPEndpoint(), "", "", "", user)
+			require.NoError(t, err)
 
-	c, err := e2emimir.NewClient(distributor.HTTPEndpoint(), "", "", "", user)
-	require.NoError(t, err)
-
-	runTestPushSeriesAndUnalignedQuery(t, c, queryFrontendAligned, queryFrontendUnaligned, user, "series_1", generateFloatSeries, generateExpectedFloats)
-	runTestPushSeriesAndUnalignedQuery(t, c, queryFrontendAligned, queryFrontendUnaligned, user, "hseries_1", generateHistogramSeries, generateExpectedHistograms)
+			runTestPushSeriesAndUnalignedQuery(t, c, queryFrontendAligned, queryFrontendUnaligned, user, "series_1", generateFloatSeries, generateExpectedFloats)
+			runTestPushSeriesAndUnalignedQuery(t, c, queryFrontendAligned, queryFrontendUnaligned, user, "hseries_1", generateHistogramSeries, generateExpectedHistograms)
+		})
+	}
 }
 
 // generateExpectedFunc defines how to generate the expected float values or histograms from the expectedVector returned with the generated series
@@ -135,6 +139,29 @@ func runTestPushSeriesAndUnalignedQuery(t *testing.T, c *e2emimir.Client, queryF
 
 		require.Equal(t, res.Type(), model.ValMatrix)
 		require.Equal(t, res.(model.Matrix), expected)
+	}
+
+}
+
+func cacheBackend(backend string) *e2e.ConcreteService {
+	if backend == cache.BackendRedis {
+		return e2ecache.NewRedis()
+	}
+
+	return e2ecache.NewMemcached()
+}
+
+func cacheConfig(backend string, service *e2e.ConcreteService) map[string]string {
+	if backend == cache.BackendRedis {
+		return map[string]string{
+			"-query-frontend.results-cache.backend":        "redis",
+			"-query-frontend.results-cache.redis.endpoint": service.NetworkEndpoint(e2ecache.RedisPort),
+		}
+	}
+
+	return map[string]string{
+		"-query-frontend.results-cache.backend":             "memcached",
+		"-query-frontend.results-cache.memcached.addresses": "dns+" + service.NetworkEndpoint(e2ecache.MemcachedPort),
 	}
 }